requirements.txt
numpy==1.21.0
pandas==1.3.0
python-binance==1.0.15
python-dotenv==0.19.0